magnet and device now using pydantic model
